Updated atoi, atoi and cmp functions

+++ libft_project/libft2/ft_memcmp.c
@@ -13,13 +13,13 @@ int		ft_memcmp(const void *s1, const void *s2, size_t n)
 	str2 = s2;
 	while (i < n)
 	{
-		ret += (int)str1[i];
+		ret += str1[i];
 		i++;
 	}
 	i = 0;
 	while (i < n)
 	{
-		ret -= (int)str2[i];
+		ret -= str2[i];
 		i++;
 	}
 	return (ret);
